cleanup react  Query implementation
