+++ src/components/ReactQuery/types.ts
@@ -1,0 +1,13 @@
+export interface Post {
+	id: number;
+	title: string;
+	body: string;
+}
+
+export interface PostsProps {
+	setPostId: (id: number) => void;
+}
+
+export interface PostProps extends PostsProps {
+	postId: number;
+}
